Move SparkMaxWrapper into the simulation directory - better organization
